requirements.txt
Flask
slack-sdk
python-dotenv
aiohttp